+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.67] - 2024-02-11
+### Added
+* `LayoutGroup3D` from `Project-LÜTRO`
+* `QuaternionExtensions` with its function `ToDirectionVector`
+* `Ballistics` from `Project-LÜTRO` to Math dir
+* `GetLastValue` and `GetCount` `EnumUtils`
+### Changed
+* `EnumUtils`'s `GetValues` to return `T[]` as oppose to `IEnumerable<T>`
+
 ## [0.0.66] - 2024-01-29
 ### Added
 * `TransformExtensions`.`TransformRotation` and `InverseTransformRotation` to convert a rotation from one transform to another